azure: adapt common provider framework
Fixes: #576
