+++ cmd/cloud-api-adaptor/main.go
@@ -13,7 +13,6 @@ import (
 	"github.com/confidential-containers/cloud-api-adaptor/cmd/cloud-api-adaptor/cloudmgr"
 	"github.com/confidential-containers/cloud-api-adaptor/pkg/adaptor"
 	"github.com/confidential-containers/cloud-api-adaptor/pkg/adaptor/hypervisor"
-	"github.com/confidential-containers/cloud-api-adaptor/pkg/adaptor/hypervisor/azure"
 	"github.com/confidential-containers/cloud-api-adaptor/pkg/adaptor/hypervisor/libvirt"
 	"github.com/confidential-containers/cloud-api-adaptor/pkg/adaptor/hypervisor/registry"
 	daemon "github.com/confidential-containers/cloud-api-adaptor/pkg/forwarder"
@@ -36,7 +35,6 @@ type networkConfig struct {
 	VXLANMinID    int
 }
 
-var azurecfg azure.Config
 var libvirtcfg libvirt.Config
 var hypcfg hypervisor.Config
 
@@ -88,33 +86,6 @@ func (cfg *daemonConfig) Setup() (cmd.Starter, error) {
 	// TODO: following lines will be removed when refactoring is done
 
 	switch os.Args[1] {
-	case "azure":
-		cmd.Parse("azure", os.Args[1:], func(flags *flag.FlagSet) {
-			flags.StringVar(&azurecfg.ClientId, "clientid", "", "Client Id, defaults to `AZURE_CLIENT_ID`")
-			flags.StringVar(&azurecfg.ClientSecret, "secret", "", "Client Secret, defaults to `AZURE_CLIENT_SECRET`")
-			flags.StringVar(&azurecfg.TenantId, "tenantid", "", "Tenant Id, defaults to `AZURE_TENANT_ID`")
-			flags.StringVar(&azurecfg.ResourceGroupName, "resourcegroup", "", "Resource Group")
-			flags.StringVar(&azurecfg.Zone, "zone", "", "Zone")
-			flags.StringVar(&azurecfg.Region, "region", "", "Region")
-			flags.StringVar(&azurecfg.SubnetId, "subnetid", "", "Network Subnet Id")
-			flags.StringVar(&azurecfg.SecurityGroupId, "securitygroupid", "", "Security Group Id")
-			flags.StringVar(&azurecfg.Size, "instance-size", "", "Instance size")
-			flags.StringVar(&azurecfg.ImageId, "imageid", "", "Image Id")
-			flags.StringVar(&azurecfg.SubscriptionId, "subscriptionid", "", "Subscription ID")
-			flags.StringVar(&azurecfg.SSHKeyPath, "ssh-key-path", "$HOME/.ssh/id_rsa.pub", "Path to SSH public key")
-			flags.StringVar(&hypcfg.SocketPath, "socket", hypervisor.DefaultSocketPath, "Unix domain socket path of remote hypervisor service")
-			flags.StringVar(&hypcfg.PodsDir, "pods-dir", hypervisor.DefaultPodsDir, "base directory for pod directories")
-			flags.StringVar(&hypcfg.HypProvider, "provider", "azure", "Hypervisor provider")
-			flags.StringVar(&hypcfg.CriSocketPath, "cri-runtime-endpoint", "", "cri runtime uds endpoint")
-			flags.StringVar(&hypcfg.PauseImage, "pause-image", "", "pause image to be used for the pods")
-			flags.StringVar(&cfg.TunnelType, "tunnel-type", podnetwork.DefaultTunnelType, "Tunnel provider")
-			flags.StringVar(&cfg.HostInterface, "host-interface", "", "Host Interface")
-			flags.IntVar(&cfg.VXLANPort, "vxlan-port", vxlan.DefaultVXLANPort, "VXLAN UDP port number (VXLAN tunnel mode only")
-			flags.IntVar(&cfg.VXLANMinID, "vxlan-min-id", vxlan.DefaultVXLANMinID, "Minimum VXLAN ID (VXLAN tunnel mode only")
-		})
-		defaultToEnv(&azurecfg.ClientId, "AZURE_CLIENT_ID")
-		defaultToEnv(&azurecfg.ClientSecret, "AZURE_CLIENT_SECRET")
-		defaultToEnv(&azurecfg.TenantId, "AZURE_TENANT_ID")
 
 	case "libvirt":
 		cmd.Parse("libvirt", os.Args[1:], func(flags *flag.FlagSet) {
@@ -143,8 +114,6 @@ func (cfg *daemonConfig) Setup() (cmd.Starter, error) {
 
 	if hypcfg.HypProvider == "libvirt" {
 		hypervisorServer = registry.NewServer(hypcfg, libvirtcfg, workerNode, daemon.DefaultListenPort)
-	} else if hypcfg.HypProvider == "azure" {
-		hypervisorServer = registry.NewServer(hypcfg, azurecfg, workerNode, daemon.DefaultListenPort)
 	}
 
 	return cmd.NewStarter(hypervisorServer), nil
