+++ pkg/adaptor/cloud/azure/types_test.go
@@ -1,4 +1,5 @@
-//go:build azure
+// (C) Copyright Confidential Containers Contributors
+// SPDX-License-Identifier: Apache-2.0
 
 package azure
